\name{MyThemeGgplot-package}
\alias{MyThemeGgplot-package}
\alias{MyThemeGgplot}
\docType{package}
\title{
\packageTitle{MyThemeGgplot}
}
\description{
\packageDescription{MyThemeGgplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MyThemeGgplot}
\packageIndices{MyThemeGgplot}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MyThemeGgplot}

Maintainer: \packageMaintainer{MyThemeGgplot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
